Upgrade Swashbuckle to 5.0.0-rc4
Upgrade Swashbuckle for dotnet core 3.0 compatibility
Move JaegerInitializer to its own file
