+++ bevrand.authenticationapi/bevrand.authenticationapi/Startup.cs
@@ -8,7 +8,6 @@ using bevrand.authenticationapi.Services;
 using bevrand.authenticationapi.Services.Interfaces;
 using OpenTracing.Util;
 using Jaeger;
-using Jaeger.Samplers;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
@@ -18,37 +17,10 @@ using Microsoft.Extensions.Logging;
 using Swashbuckle.AspNetCore.Swagger;
 using Microsoft.AspNetCore.Mvc.NewtonsoftJson;
 using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace bevrand.authenticationapi
 {
-    /// <summary>
-    /// Initial configuration for Jaeger Client
-    /// </summary>
-    public static class JaegerInitializer
-    {
-        /// <summary>
-        /// Get configured instance of Jaeger Client 
-        /// </summary>
-        /// <param name="serviceName">Name of service for which Tracer is configured</param>
-        /// <param name="loggerFactory">Logger factory which should be used by Jaeger Client</param>
-        /// <returns> Jaeger Client instance</returns>
-        public static Tracer Init(string serviceName, ILoggerFactory loggerFactory)
-        {
-            var samplerConfiguration = new Configuration.SamplerConfiguration(loggerFactory)
-                .WithType(ConstSampler.Type)
-                .WithParam(1);
-
-            var reporterConfiguration = new Configuration.ReporterConfiguration(loggerFactory)
-                .WithSender(Configuration.SenderConfiguration.FromEnv(loggerFactory))
-                .WithLogSpans(true);
-
-            return (Tracer)new Configuration(serviceName, loggerFactory)
-                .WithSampler(samplerConfiguration)
-                .WithReporter(reporterConfiguration)
-                .GetTracer();
-        }
-    }
-    
     public class Startup
     {
         private static readonly ILoggerFactory JaegerLoggerFactory = LoggerFactory.Create(builder =>
@@ -70,9 +42,7 @@ namespace bevrand.authenticationapi
             Configuration = builder.Build();
         }
         
-
         public IConfiguration Configuration { get; }
-
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -99,11 +69,10 @@ namespace bevrand.authenticationapi
                 var commentsFileName = Assembly.GetExecutingAssembly().GetName().Name + ".xml";
                 var commentsFile = Path.Combine(baseDirectory, commentsFileName);
                 
-                c.SwaggerDoc("v1", new Info { Title = "AuthenticationApi", Version = "v1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "AuthenticationApi", Version = "v1" });
                 c.IncludeXmlComments(commentsFile);
             });
         }
-        
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
@@ -112,6 +81,8 @@ namespace bevrand.authenticationapi
             {
                 app.UseDeveloperExceptionPage();
             }
+
+            app.UseStaticFiles();
 
             app.UseRouting();
